requirements.txt
playwright==1.49.1
python-telegram-bot==21.9
python-dotenv==1.0.1
apscheduler==3.10.4
cryptography==44.0.2
